+++ src/lune/lua/mod.rs
@@ -9,6 +9,5 @@ pub mod serde;
 pub mod stdio;
 pub mod table;
 pub mod task;
-pub mod time;
 
 pub use create::create as create_lune_lua;
